Move examples and symbol guide.
