+++ reference-guide/src/main/scala/docs/reference/Err.scala
@@ -1,0 +1,3 @@
+package docs.reference
+
+object Err extends Throwable("Err")
