+++ reference-guide/src/main/scala/docs/reference/bracket.scala
