+++ reference-guide/src/main/scala/docs/reference/drop.scala
