+++ reference-guide/src/main/scala/docs/reference/errors.scala
