+++ reference-guide/src/main/scala/docs/reference/evalMap.scala
