+++ reference-guide/src/main/scala/docs/reference/filter.scala
