+++ reference-guide/src/main/scala/docs/reference/flatMap.scala
