+++ reference-guide/src/main/scala/docs/reference/merge.scala
