+++ reference-guide/src/main/scala/docs/reference/parEvalMap.scala
