+++ reference-guide/src/main/scala/docs/reference/scan.scala
