+++ reference-guide/src/main/scala/docs/reference/take.scala
