+++ symbol-guide/src/main/scala/drawing/Symbols.scala
